requirements.txt
py-cord==2.4.1
pymongo==4.6.1
python-dotenv==1.0.0
pytz==2022.7.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
